+++ CHANGELOG.md
@@ -77,6 +77,7 @@
 - Setup.py fixed for Python 3.11.
 - Binance parser: added re-named operations for Statements. ([#292](https://github.com/BittyTax/BittyTax/issues/292))
 - OKX parser: added new export formats. ([#293](https://github.com/BittyTax/BittyTax/issues/293))
+- Exodus parser: added staking export.
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.